Testing/1.4.15

+++ test_package/conanfile.py
@@ -18,7 +18,9 @@ class TestPackageConan(ConanFile):
 
     def test(self):
         with tools.environment_append(RunEnvironment(self).vars):
-            process = subprocess.Popen(["mosquitto"])
+            process = None
+            if self.options["mosquitto:with_binaries"] == True:
+                process = subprocess.Popen(["mosquitto"])
             bin_path = os.path.join("bin", "test_package")
             if self.settings.os == "Windows":
                 self.run(bin_path)
@@ -26,4 +28,5 @@ class TestPackageConan(ConanFile):
                 self.run("DYLD_LIBRARY_PATH=%s %s" % (os.environ.get('DYLD_LIBRARY_PATH', ''), bin_path))
             else:
                 self.run("LD_LIBRARY_PATH=%s %s" % (os.environ.get('LD_LIBRARY_PATH', ''), bin_path))
-            process.kill()
+            if self.options["mosquitto:with_binaries"] == True:
+                process.kill()
